requirements.txt
# python==3.11
openai
numpy
pandas
python-dotenv
pyyml
tenacity
tiktoken
tqdm
gensim
jsonlines
astunparse
pyarrow
google-generativeai
accelerate
groq
matplotlib
requests
